requirements.txt
aiohttp==2.2.2
async-timeout==1.2.1
asyncio==3.4.3
awesome-slugify==1.6.5
beautifulsoup4==4.6.0
bs4==0.0.1
chardet==3.0.4
multidict==3.1.0
regex==2017.6.23
Unidecode==0.4.21
yarl==0.11.0
